+++ app/src/views/AccountDetails/index.tsx
@@ -38,7 +38,6 @@ import i18n from './i18n.json';
 
 const TOKEN_PAGE_SIZE = 12;
 
-/** @knipignore */
 // eslint-disable-next-line import/prefer-default-export
 export function Component() {
     const strings = useTranslation(i18n);