+++ app/src/views/AccountMyFormsDref/index.tsx
@@ -18,7 +18,6 @@ import DownloadImportTemplateButton from './DownloadImportTemplateButton';
 import i18n from './i18n.json';
 import styles from './styles.module.css';
 
-/** @knipignore */
 // eslint-disable-next-line import/prefer-default-export
 export function Component() {
     const [currentView, setCurrentView] = useState<'active' | 'completed'>('active');