+++ app/src/views/AllSurgeAlerts/index.tsx
@@ -71,7 +71,6 @@ function getMolnixKeywords(molnixTags: SurgeListItem['molnix_tags']) {
         .join(', ');
 }
 
-/** @knipignore */
 // eslint-disable-next-line import/prefer-default-export
 export function Component() {
     const strings = useTranslation(i18n);