+++ app/src/views/RecoverAccount/index.tsx
@@ -44,7 +44,6 @@ const formSchema: FormSchema = {
     }),
 };
 
-/** @knipignore */
 // eslint-disable-next-line import/prefer-default-export
 export function Component() {
     const strings = useTranslation(i18n);