+++ app/src/views/SurgeCatalogue/index.tsx
@@ -40,7 +40,6 @@ import TabPage from '#components/TabPage';
 import i18n from './i18n.json';
 import styles from './styles.module.css';
 
-/** @knipignore */
 // eslint-disable-next-line import/prefer-default-export
 export function Component() {
     const strings = useTranslation(i18n);