+++ app/src/views/SurgeOperationalToolbox/index.tsx
@@ -431,7 +431,6 @@ const operationTimelineContent: OperationTimelineContent[] = [
     },
 ];
 
-/** @knipignore */
 // eslint-disable-next-line import/prefer-default-export
 export function Component() {
     const strings = useTranslation(i18n);