+++ app/src/views/ThreeWActivityDetail/index.tsx
@@ -40,7 +40,6 @@ function simplifiedKeySelector(item: ActivityItem) {
     return item.id;
 }
 
-/** @knipignore */
 // eslint-disable-next-line import/prefer-default-export
 export function Component() {
     const strings = useTranslation(i18n);