+++ packages/ui/src/components/TabListLayout/index.tsx
@@ -11,7 +11,7 @@ import { SpacingType } from '#utils/style';
 import styles from './styles.module.css';
 
 export interface Props extends React.HTMLProps<HTMLDivElement> {
-    elementRef?: React.RefObject<HTMLDivElement>;
+    elementRef?: React.RefObject<HTMLDivElement | null>;
     className?: string;
     spacing?: SpacingType;
     children: React.ReactNode;
@@ -61,6 +61,7 @@ function TabListLayout(props: Props) {
                     role="tablist"
                     spacing={styleVariant === 'vertical' ? 'none' : spacing}
                     withSpacingOpticalCorrection={styleVariant === 'vertical-compact'}
+                    spacingOffset={styleVariant === 'vertical-compact' ? -2 : 0}
                 >
                     {children}
                 </ListView>
@@ -75,6 +76,7 @@ function TabListLayout(props: Props) {
                     role="tablist"
                     withWrap={styleVariant === 'nav'}
                     withSpacingOpticalCorrection
+                    withStartAlignment={styleVariant === 'step'}
                 >
                     {children}
                 </ListView>