+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,130 @@
 # @ifrc-go/ui
 
+## 2.2.0
+
+### Minor Changes
+
+- 4b0cd41: Improve various components
+
+  - BlockView: Add support for overflow
+  - ButtonLayout: Add 'xs' text size
+  - Container
+    - Fix content overflow not working in absence of withPadding
+    - Add support for overflow
+  - Description: Add 'xs' text size
+  - DropdownMenu: Add option to remove popup padding
+  - ListView
+    - Add support for different sidebar sizes
+    - Add support for overflow
+    - Add suport for grow
+  - TabLayout
+    - Improve spacing for vertical compact variant
+    - Add an active indicator icon for vertical compact variant
+    - Add visual feedback
+  - TabListLayout: Improve spacing for vertical compact variant
+  - TabPanel: Add support for contents only display
+  - Update default line height to 1.5
+  - Update optical correction factor to match the new line height
+
+### Patch Changes
+
+- cf73a97: Update DateInput to support month input
+
+## 2.2.0-beta.1
+
+### Patch Changes
+
+- cf73a97: Update DateInput to support month input
+
+## 2.2.0-beta.0
+
+### Minor Changes
+
+- 4b0cd41: Improve various components
+
+  - BlockView: Add support for overflow
+  - ButtonLayout: Add 'xs' text size
+  - Container
+    - Fix content overflow not working in absence of withPadding
+    - Add support for overflow
+  - Description: Add 'xs' text size
+  - DropdownMenu: Add option to remove popup padding
+  - ListView
+    - Add support for different sidebar sizes
+    - Add support for overflow
+    - Add suport for grow
+  - TabLayout
+    - Improve spacing for vertical compact variant
+    - Add an active indicator icon for vertical compact variant
+    - Add visual feedback
+  - TabListLayout: Improve spacing for vertical compact variant
+  - TabPanel: Add support for contents only display
+  - Update default line height to 1.5
+  - Update optical correction factor to match the new line height
+
+## 2.1.0
+
+### Minor Changes
+
+- e619fc8: Improve various UI components
+
+  - Add form variant in Heading
+  - Add option to show border in container
+  - Restructure radio icon size
+  - Improve info popup icon size
+  - Update styling of TabLayout
+    - Update 'step' variant TabLayout to match the designs
+    - Add 'form' variant to Container
+    - Fix styling of info popup
+    - Update styling of InputLabel
+
+### Patch Changes
+
+- 22038c4: Remove pointer events from Dropdown button children
+- bd93064: Update TabLayout and Message component
+
+  - Add foreground color as background in message
+  - Increase max width of step tab
+  - Add overflow wrap to step tab
+  - Improve spacing of step & vertical compact tab
+
+- 52aefd2: Fix RadioInput
+
+  - Update double onChange callback
+
+## 2.1.0-beta.3
+
+### Patch Changes
+
+- bd93064: Update TabLayout and Message component
+
+  - Add foreground color as background in message
+  - Increase max width of step tab
+  - Add overflow wrap to step tab
+  - Improve spacing of step & vertical compact tab
+
+## 2.1.0-beta.2
+
+### Minor Changes
+
+- e619fc8: Improve various UI components
+
+  - Add form variant in Heading
+  - Add option to show border in container
+  - Restructure radio icon size
+  - Improve info popup icon size
+  - Update styling of TabLayout
+    - Update 'step' variant TabLayout to match the designs
+    - Add 'form' variant to Container
+    - Fix styling of info popup
+    - Update styling of InputLabel
+
+## 2.0.1-beta.1
+
+### Patch Changes
+
+- 22038c4: Remove pointer events from Dropdown button children
+
 ## 2.0.1-beta.0
 
 ### Patch Changes